js/xpconnect/tests/mochitest/test_bug912322.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=912322
-->
<head>
  <meta charset="utf-8">
  <title>Test for Bug 912322</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
  <script type="application/javascript">

  /** Test that XBL functions aren't exposed to the web. **/
  funs = ['getAnonymousNodes', 'getAnonymousElementByAttribute',
          'getBindingParent', 'loadBindingDocument'];
  for (var f of funs) {
    ok(!(f in document), f + " should not be available to content");
    ok(f in SpecialPowers.wrap(document), f + " should be available to chrome via Xray");
  }




  </script>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=912322">Mozilla Bug 912322</a>
<p id="display"></p>
<div id="content" style="display: none">

</div>
<pre id="test">
</pre>
</body>
</html>
